# homework4/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthResponse:
    properties:
      token:
        type: string
    type: object
  dto.CreateCommentRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  dto.CreatePostRequest:
    properties:
      content:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - content
    - title
    type: object
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.RegisterRequest:
    properties:
      password:
        maxLength: 128
        minLength: 6
        type: string
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  dto.UpdatePostRequest:
    properties:
      content:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - content
    - title
    type: object
  model.Comment:
    properties:
      author:
        $ref: '#/definitions/model.User'
      authorId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      postId:
        type: integer
      updatedAt:
        type: string
    type: object
  model.Post:
    properties:
      author:
        $ref: '#/definitions/model.User'
      authorId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: http://localhost:8080
info:
  contact: {}
  description: 博客项目API接口文档
  title: 项目接口
  version: 1.0 版本
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户登录
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - Auth
  /posts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
      summary: 获取文章列表
      tags:
      - Post
    post:
      consumes:
      - application/json
      parameters:
      - description: create post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Post'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 新增文章
      tags:
      - Post
  /posts/{id}:
    delete:
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除文章
      tags:
      - Post
    get:
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取文章详情
      tags:
      - Post
    put:
      consumes:
      - application/json
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: integer
      - description: update post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新文章内容
      tags:
      - Post
  /posts/{id}/comments:
    get:
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Comment'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取指定文章的评论
      tags:
      - Comment
    post:
      consumes:
      - application/json
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: integer
      - description: create comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Comment'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 新增评论
      tags:
      - Comment
swagger: "2.0"